feat: 1차 모듈구현(인풋 에러메시지 표시파트)
