+++ script/signin.js
@@ -1,11 +1,15 @@
 "use strict";
+// import 할 때 한 번에 가져오기
+import {
+  emailErrHandler, // 이벤트헨들러
+  pwErrHandler, // 이벤트헨들러
+  emailInput, // 상수
+  pwInput, // 상수
+} from "./module/signForm/inputErrMsg.js";
 
 document.addEventListener("DOMContentLoaded", function () {
   // =================================================================
-  const emailInput = document.querySelector("#email");
-  const pwInput = document.querySelector("#password");
-  const emailErrorMsg = document.querySelector("#email_error");
-  const pwErrorMsg = document.querySelector("#pw_error");
+
   const pwIcon = document.querySelector(".password_eyes");
   const loginBtn = document.querySelector("#login_btn");
 
@@ -31,50 +35,11 @@ document.addEventListener("DOMContentLoaded", function () {
     e.keyCode === 13 && loginHandler(e);
   });
 
-  // 에러이벤트 표시용 함수
-  function showError(errorMsg, input, message) {
-    errorMsg.classList.remove("done");
-    errorMsg.classList.add("show");
-    errorMsg.innerHTML = message;
-    input.classList.add("errorInput_style");
-  }
-  // 에러이벤트 숨김용 함수
-  function doneError(errorMsg, input) {
-    errorMsg.classList.remove("show");
-    errorMsg.classList.add("done");
-    input.classList.remove("errorInput_style");
-  }
-
   //이메일인풋 에러핸들러
-  emailInput.addEventListener("blur", () => {
-    if (emailInput.value === "") {
-      showError(emailErrorMsg, emailInput, "이메일 주소를 입력하세요.");
-    } else if (!/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(emailInput.value)) {
-      showError(emailErrorMsg, emailInput, "올바른 이메일 주소가 아닙니다.");
-    } else if (emailInput.value === "test@codeit.com") {
-      showError(emailErrorMsg, emailInput, "이미 사용 중인 이메일입니다.");
-    } else {
-      doneError(emailErrorMsg, emailInput);
-    }
-  });
+  emailInput.addEventListener("blur", emailErrHandler);
   //패스워드인풋 에러핸들러
-  pwInput.addEventListener("blur", () => {
-    if (pwInput.value === "") {
-      showError(pwErrorMsg, pwInput, "비밀번호를 입력해주세요.");
-    } else if (
-      pwInput.value.length < 8 ||
-      /^[a-zA-Z]+$/.test(pwInput.value) ||
-      /^\d+$/.test(pwInput.value)
-    ) {
-      showError(
-        pwErrorMsg,
-        pwInput,
-        "비밀번호는 8자 이상이어야 하며, 문자와 숫자를 혼합해야 합니다."
-      );
-    } else {
-      doneError(pwErrorMsg, pwInput);
-    }
-  });
+  pwInput.addEventListener("blur", pwErrHandler);
+
   //아이콘 변경용 핸들러
   pwIcon.addEventListener("click", () => {
     if (pwVisibleToggle) {
