+++ script/signup.js
@@ -1,13 +1,16 @@
 "use strict";
+import {
+  emailErrHandler,
+  pwErrHandler,
+  pwCkdErrHandler,
+  emailInput,
+  pwInput,
+  pwCheckInput,
+} from "./module/signForm/inputErrMsg.js";
 
 document.addEventListener("DOMContentLoaded", function () {
   // =================================================================
-  const emailInput = document.querySelector("#email");
-  const pwInput = document.querySelector("#password");
-  const pwCheckInput = document.querySelector("#password_check");
-  const emailErrorMsg = document.querySelector("#email_error");
-  const pwErrorMsg = document.querySelector("#pw_error");
-  const pwCheckErrorMsg = document.querySelector("#pwCheck_error");
+
   const icons = document.querySelectorAll(".password_eyes");
   const signUpBtn = document.querySelector("#signup_btn");
   const pwIcon = icons[0];
@@ -32,43 +35,11 @@ document.addEventListener("DOMContentLoaded", function () {
   }
 
   //이메일인풋 에러핸들러
-  emailInput.addEventListener("blur", () => {
-    if (emailInput.value === "") {
-      showError(emailErrorMsg, emailInput, "이메일 주소를 입력하세요.");
-    } else if (!/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(emailInput.value)) {
-      showError(emailErrorMsg, emailInput, "올바른 이메일 주소가 아닙니다.");
-    } else if (emailInput.value === "test@codeit.com") {
-      showError(emailErrorMsg, emailInput, "이미 사용 중인 이메일입니다.");
-    } else {
-      doneError(emailErrorMsg, emailInput);
-    }
-  });
+  emailInput.addEventListener("blur", emailErrHandler);
   //패스워드인풋 에러핸들러
-  pwInput.addEventListener("blur", () => {
-    if (pwInput.value === "") {
-      showError(pwErrorMsg, pwInput, "비밀번호를 입력해주세요.");
-    } else if (
-      pwInput.value.length < 8 ||
-      /^[a-zA-Z]+$/.test(pwInput.value) ||
-      /^\d+$/.test(pwInput.value)
-    ) {
-      showError(
-        pwErrorMsg,
-        pwInput,
-        "비밀번호는 8자 이상이어야 하며, 문자와 숫자를 혼합해야 합니다."
-      );
-    } else {
-      doneError(pwErrorMsg, pwInput);
-    }
-  });
+  pwInput.addEventListener("blur", pwErrHandler);
   //패스워드체크(재확인) 에러핸들러
-  pwCheckInput.addEventListener("blur", () => {
-    if (pwInput.value !== pwCheckInput.value) {
-      showError(pwCheckErrorMsg, pwCheckInput, "비밀번호가 일치하지 않아요.");
-    } else {
-      doneError(pwCheckErrorMsg, pwCheckInput);
-    }
-  });
+  pwCheckInput.addEventListener("blur", pwCkdErrHandler);
   //아이콘 변경용 핸들러
   pwIcon.addEventListener("click", () => {
     if (pwVisibleToggle) {
