+++ signin.html
@@ -13,7 +13,7 @@
       crossorigin
       href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.8/dist/web/variable/pretendardvariable-dynamic-subset.css"
     />
-    <script defer src="script/signin.js"></script>
+    <script type="module" defer src="script/signin.js"></script>
   </head>
   <body>
     <!-- ================================== -->
